server/requirements.txt
Flask==2.2.2
Flask_Cors==3.0.10
Flask_JWT_Extended==4.4.4
pymongo==4.3.3
python-dotenv==1.0.0


attrdict-2.0.1 
chardet-3.0.4 
eth-abi-1.3.0
eth-account-9.3.0
eth-hash-0.2.0
eth-keyfile-0.5.1
eth-keys-0.2.1
eth-rlp-0.1.2 
eth-typing-2.1.0
eth-utils-1.4.1
hexbytes-0.1.0
idna-2.8
lru-dict-1.1.6
parsimonious-0.8.1
pycryptodome-3.8.0
requests-2.21.0
rlp-1.1.0
six-1.12.0 
urllib3-1.24.1
web3-4.9.1
